Rename server -> backend, client -> frontend
