+++ src/backend/sql/count.js
